% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{adjust_times}
\alias{adjust_times}
\title{Create Time column}
\usage{
adjust_times(dt, sfreq, meas_start, samp_num)
}
\arguments{
\item{dt}{data.frame to use}

\item{sfreq}{sampling frequency}

\item{meas_start}{start of measurement}

\item{samp_num}{column containing sample numbers}
}
\value{
data.frame
}
\description{
Create Time column
}
