% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{asym_model}
\alias{asym_model}
\title{Self Starting Asymptotic Regression}
\usage{
asym_model(df)
}
\arguments{
\item{df}{data frame}
}
\value{
list of coefficients from regression if successful, NA if unable to run
}
\description{
Self Starting Asymptotic Regression
}
