% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{avg_type}
\alias{avg_type}
\title{Average values across data type}
\usage{
avg_type(dt, col = character(), groups = character())
}
\arguments{
\item{dt}{data.frame}

\item{col}{character, name of column to calculate on}

\item{groups}{columns to group by}
}
\value{
data.frame
}
\description{
Average values across data type
}
