% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{bads}
\alias{bads}
\title{Acessor for \code{bads}}
\usage{
bads(x)
}
\arguments{
\item{x}{object}
}
\value{
value if \code{bads} exists in the object, NULL otherwise
}
\description{
Acessor for \code{bads}
}
