% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_info.R
\name{create_info}
\alias{create_info}
\title{Create an Info object}
\usage{
create_info(
  col_names = character(),
  col_types = "misc",
  device_type = NULL,
  device_model = NULL,
  device_serial = NULL,
  device_site = NULL,
  device_info = NULL,
  subj_id = NULL,
  subj_sex = NULL,
  subj_age = NULL,
  subj_height = NULL,
  subj_weight = NULL,
  subj_info = NULL,
  desc = NULL,
  meas_date = NULL,
  meas_id = NULL,
  meas_start = NULL,
  samps = NULL,
  sfreq = NULL,
  bads = NULL
)
}
\arguments{
\item{col_names}{A character vector of names of data columns}

\item{col_types}{A character vector of data types. Defaults to "misc".}

\item{device_type}{Device type that collected the data.}

\item{device_model}{Device model name.}

\item{device_serial}{Device serial identifier.}

\item{device_site}{Device site.}

\item{device_info}{A list of device information.}

\item{subj_id}{subject identifier}

\item{subj_sex}{subject sex}

\item{subj_age}{subject age}

\item{subj_height}{subject height}

\item{subj_weight}{subject weight}

\item{subj_info}{subject information as a list}

\item{desc}{measurement description}

\item{meas_date}{datetime of the measurement}

\item{meas_id}{measurement id}

\item{meas_start}{time to use as the start of the testing}

\item{samps}{name of column containing sample numbers}

\item{sfreq}{sampling frequency in Hz}

\item{bads}{list of channels that have bad signal}
}
\value{
Info object containing all provided parameters. The Info object
inherits from list, so will behave as a list.
}
\description{
Create \code{info} object to record measurement metadata. This object behaves
like a list. It contains all metadata available for the measurement.
The only required input is \code{col_names}. Even if \code{col_names} is the only
user-supplied input, the \code{info} object will have empty slots for all possible
values.
}
\details{
Implementation is based on the info class for MNE python.

\code{col_names}is the only required argument. \code{col_types} defaults to "misc"
for all cols without other input. If both \code{col_names} and \code{col_types} are
provided, the length of \code{col_names} must match the length of \code{col_types} if
\code{col_types} is longer than 1. If \code{col_types} is of length 1, then all cols
will be that type.
}
