% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_sections}
\alias{create_sections}
\title{Create on/off Kinetics sections}
\usage{
create_sections(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
data.frame
}
\description{
Create on/off Kinetics sections
}
