% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.R
\name{crop}
\alias{crop}
\title{Remove rows from data}
\usage{
crop(data, start, end, col)
}
\arguments{
\item{data}{data.frame}

\item{start}{A numeric. If `col` is provided, rows with value less than
`start` in column `col` are removed.  If `col` is not provided, rows
with row number less than `start` are removed.
Defaults to NULL.}

\item{end}{A numeric. If `col` is provided, rows with value greater than
`end` in column `end` are removed.  If `col` is not provided, rows with
row number greater than `end` are removed.
Defaults to NULL.}

\item{col}{A character name of column to use to evaluate `start` and `end`.
Defaults to NULL.

If `start`,`end`, and `col` are NULL, the original data will be returned
with a message that no rows were cropped.

`start` and `end` are not inclusive. If `start` = 1, row 1 will not be removed.
To remove row 1, `start` must be 2.}
}
\value{
data.frame
}
\description{
Delete rows from data. If `col` is provided, works like
`dplyr::filter` and filters values in column `col` that are
greater than `start` and less than `end`.
If `col` is not provided, works like subsetting based on row numbers.
Rows with row numbers less than `start` and greater than `end` are removed from
the dataset.
}
