% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{device_model<-}
\alias{device_model<-}
\title{Set the value of \code{device_model}}
\usage{
device_model(x) <- value
}
\arguments{
\item{x}{object}

\item{value}{new value}
}
\value{
modified object
}
\description{
Set the value of \code{device_model}
}
