% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{device_model}
\alias{device_model}
\title{Acessor for \code{device_model}}
\usage{
device_model(x)
}
\arguments{
\item{x}{object}
}
\value{
value if \code{device_model} exists in the object, NULL otherwise
}
\description{
Acessor for \code{device_model}
}
