% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{device_serial}
\alias{device_serial}
\title{Acessor for `device_serial`}
\usage{
device_serial(x)
}
\arguments{
\item{x}{object}
}
\value{
value if `device_serial` exists in the object, NULL otherwise
}
\description{
Acessor for `device_serial`
}
