% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{device_site<-}
\alias{device_site<-}
\title{Set the value of \code{device_site}}
\usage{
device_site(x) <- value
}
\arguments{
\item{x}{object}

\item{value}{new value}
}
\value{
modified object
}
\description{
Set the value of \code{device_site}
}
