% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{device_type<-}
\alias{device_type<-}
\title{Set the value of `device_type`}
\usage{
device_type(x) <- value
}
\arguments{
\item{x}{object}

\item{value}{new value}
}
\value{
modified object
}
\description{
Set the value of `device_type`
}
