% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{device_type}
\alias{device_type}
\title{Acessor for `device_type`}
\usage{
device_type(x)
}
\arguments{
\item{x}{object}
}
\value{
value if `device_type` exists in the object, NULL otherwise
}
\description{
Acessor for `device_type`
}
