% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nirs.R
\name{extract_ext}
\alias{extract_ext}
\title{Get the file extension type}
\usage{
extract_ext(file)
}
\arguments{
\item{file}{name of the file}
}
\value{
file extension type
* csv
* xlsx
* txt
}
\description{
Get the file extension type
}
