% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{first_samp<-}
\alias{first_samp<-}
\title{Setter for \code{first_samp}}
\usage{
first_samp(x) <- value
}
\arguments{
\item{x}{object}

\item{value}{value to set as \code{first_samp}}
}
\value{
modified object
}
\description{
Setter for \code{first_samp}
}
