% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{first_samp}
\alias{first_samp}
\title{Accessor for \code{first_samp}}
\usage{
first_samp(x)
}
\arguments{
\item{x}{object}
}
\value{
value of \code{first_samp}
}
\description{
Accessor for \code{first_samp}
}
