% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggook}
\alias{ggook}
\title{Wrapper around ggplot functions to plot on/off kinetics}
\usage{
ggook(data, x, y)
}
\arguments{
\item{data}{data frame to use}

\item{x}{limits}

\item{y}{limits}
}
\description{
Wrapper around ggplot functions to plot on/off kinetics
}
