% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{group_nest_dt}
\alias{group_nest_dt}
\title{Nest data.table by Group}
\usage{
group_nest_dt(dt, ..., .key = "data")
}
\arguments{
\item{dt}{data.table}

\item{...}{other arguments}

\item{.key}{key to group by}
}
\value{
nested data.table
}
\description{
Nest data.table by Group
}
\details{
function created by Tyson Barrett
}
