% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onoff_kinetics_module.R
\name{interact_rawdata}
\alias{interact_rawdata}
\title{Plotting gadget for rawdata object}
\usage{
interact_rawdata(rawdata)
}
\arguments{
\item{rawdata}{rawdata object}
}
\value{
rawdata object
}
\description{
Plotting gadget for rawdata object
}
