% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{manual_color}
\alias{manual_color}
\title{manual color scale for ggplot}
\usage{
manual_color(col_names)
}
\arguments{
\item{col_names}{character vector of unique column names in the dataset}
}
\value{
named vector of colors to use with each col_name
}
\description{
manual color scale for ggplot
}
