% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{meas_date<-}
\alias{meas_date<-}
\title{Set the value of \code{meas_date}}
\usage{
meas_date(x) <- value

meas_date(x) <- value
}
\arguments{
\item{x}{object}

\item{value}{new value}
}
\value{
modified object

modified object
}
\description{
Set the value of \code{meas_date}

Set the value of \code{meas_date}
}
