% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{meas_end<-}
\alias{meas_end<-}
\title{Set the value of \code{meas_end}}
\usage{
meas_end(x) <- value
}
\arguments{
\item{x}{object}

\item{value}{new value}
}
\value{
modified object
}
\description{
Set the value of \code{meas_end}
}
