% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{meas_end}
\alias{meas_end}
\title{Acessor for `meas_end`}
\usage{
meas_end(x)
}
\arguments{
\item{x}{object}
}
\value{
value if `meas_end` exists in the object, NULL otherwise
}
\description{
Acessor for `meas_end`
}
