% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{meas_id<-}
\alias{meas_id<-}
\title{Set the value of `meas_id`}
\usage{
meas_id(x) <- value
}
\arguments{
\item{x}{object}

\item{value}{new value}
}
\value{
modified object
}
\description{
Set the value of `meas_id`
}
