% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{meas_id}
\alias{meas_id}
\title{Acessor for \code{meas_id}}
\usage{
meas_id(x)
}
\arguments{
\item{x}{object}
}
\value{
value if \code{meas_id} exists in the object, NULL otherwise
}
\description{
Acessor for \code{meas_id}
}
