% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{meas_start}
\alias{meas_start}
\title{Acessor for `meas_start`}
\usage{
meas_start(x)
}
\arguments{
\item{x}{object}
}
\value{
value of `meas_start` if it exists, otherwise NULL
}
\description{
Acessor for `meas_start`
}
