% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_analyzeddata.R
\name{new_analyzed_data}
\alias{new_analyzed_data}
\title{Analyzed Data Class Constructor}
\usage{
new_analyzed_data(data, info, coefs, ..., class = character())
}
\arguments{
\item{data}{data object}

\item{info}{info object}

\item{coefs}{coefficients from models}

\item{...}{other variables for children of class}

\item{class}{for children of class}
}
\value{
analyzed object
}
\description{
Analyzed Data Class Constructor
}
