% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_models.R
\name{new_models}
\alias{new_models}
\title{Models Class Constructor}
\usage{
new_models(coefs, predicted, ..., class = character())
}
\arguments{
\item{models}{models}
}
\value{
model object
}
\description{
Models Class Constructor
}
