% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onoff_kinetics_module.R
\name{plot_rawdata}
\alias{plot_rawdata}
\title{Plotting gadget for rawdata object}
\usage{
plot_rawdata(rawdata)
}
\arguments{
\item{rawdata}{rawdata object}
}
\value{
TODO
}
\description{
Plotting gadget for rawdata object
}
