% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_regressions}
\alias{plot_regressions}
\title{Plot Regressions on top of actual values}
\usage{
plot_regressions(models)
}
\arguments{
\item{models}{analyzed object}
}
\value{
ggplot object
}
\description{
Plot Regressions on top of actual values
}
