% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prepare_for_modeling}
\alias{prepare_for_modeling}
\title{Prepare data for regressions}
\usage{
prepare_for_modeling(rawdata)
}
\arguments{
\item{rawdata}{rawdata object}
}
\value{
data.frame
}
\description{
Prepare data for regressions
}
