% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{prepare_for_ook_plot}
\alias{prepare_for_ook_plot}
\title{Manipulate data frame for use in plotting functions}
\usage{
prepare_for_ook_plot(rawdata, type = NULL)
}
\arguments{
\item{rawdata}{rawdata object}

\item{type}{type of data to display

internal function for use in ook_plot and plot_rawdata}
}
\description{
Manipulate data frame for use in plotting functions
}
