% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{prepare_shiny_plot}
\alias{prepare_shiny_plot}
\title{Prepare data to pass into shiny plotting function}
\usage{
prepare_shiny_plot(
  data,
  sfreq,
  meas_start,
  meas_end,
  bads,
  removed,
  col_types,
  col_names,
  type = NULL
)
}
\arguments{
\item{data}{data.frame}

\item{sfreq}{sampling frequency}

\item{meas_start}{start of measurement}

\item{meas_end}{end of measurement}

\item{bads}{character vector of column names marked as "bad"}

\item{removed}{integer vector of sample numbers to remove from the dataset}

\item{col_types}{character vector of column types}

\item{col_names}{character vector of column names}

\item{type}{type of signal to display on the graph. Defaults to NULL and displays all types.}
}
\value{
plotdata
}
\description{
Prepare data to pass into shiny plotting function
}
