% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rawdata.R
\name{rawdata}
\alias{rawdata}
\title{Rawdata Class Helper}
\usage{
rawdata(
  data = data.frame(),
  col_names = character(),
  col_types = "misc",
  device_info = NULL,
  subj_info = NULL,
  meas_date = NULL,
  meas_start = NULL,
  sfreq = NULL,
  samps = "samp_num",
  bads = NULL
)
}
\arguments{
\item{data}{data frame of raw data}

\item{col_names}{character vector of column names}

\item{col_types}{character vector of column types}

\item{device_info}{list of device information}

\item{subj_info}{list of subject information}

\item{meas_date}{date of measurement}

\item{meas_start}{time to use as the start of the test}

\item{sfreq}{sample frequency}

\item{samps}{name of column of sample numbers}

\item{bads}{list of columns that were of bad quality and should not be used in analysis}
}
\value{
rawdata
}
\description{
Rawdata Class Helper
}
