% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nirs.R
\name{read_nirs_example}
\alias{read_nirs_example}
\title{Retrieve example data sets}
\usage{
read_nirs_example(path = NULL)
}
\arguments{
\item{path}{path to example data sets}
}
\value{
file path
}
\description{
Retrieve example data sets
}
\examples{
read_nirs_example()
read_nirs_example("on_off_kinetics.txt")

}
