% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_bad.R
\name{remove_bad}
\alias{remove_bad}
\title{Remove bad columns from the dataset}
\usage{
remove_bad(dt, bads)
}
\arguments{
\item{dt}{data.frame of data}

\item{bads}{columns to remove}
}
\value{
good data
}
\description{
Remove bad columns from the dataset
}
