% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{removed<-}
\alias{removed<-}
\title{Set the value of \code{removed}}
\usage{
removed(x) <- value
}
\arguments{
\item{x}{object}

\item{value}{new value}
}
\value{
modified object
}
\description{
Set the value of \code{removed}
}
