% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{removed}
\alias{removed}
\title{Acessor for `removed`}
\usage{
removed(x)
}
\arguments{
\item{x}{object}
}
\value{
value if `removed` exists in the object, NULL otherwise
}
\description{
Acessor for `removed`
}
