% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{run_models}
\alias{run_models}
\title{Run Regressions on On/Off Kinetics Data}
\usage{
run_models(rawdata)
}
\arguments{
\item{rawdata}{rawdata object}
}
\value{
dataframe
}
\description{
Run Regressions on On/Off Kinetics Data
}
