% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{select_types}
\alias{select_types}
\title{Filter by selected column type/s}
\usage{
select_types(dt, type = NULL)
}
\arguments{
\item{dt}{data.frame}

\item{type}{character vector of column type to include.}
}
\value{
data
}
\description{
Filter by selected column type/s
}
