% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{sfreq<-}
\alias{sfreq<-}
\title{Set the value of `sfreq`}
\usage{
sfreq(x) <- value
}
\arguments{
\item{x}{object}

\item{value}{new value}
}
\value{
modified object
}
\description{
Set the value of `sfreq`
}
