% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{sfreq}
\alias{sfreq}
\title{Acessor for \code{sfreq}}
\usage{
sfreq(x)
}
\arguments{
\item{x}{object}
}
\value{
value if \code{sfreq} exists in the object, NULL otherwise
}
\description{
Acessor for \code{sfreq}
}
