% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{steady_states}
\alias{steady_states}
\title{Calculate steady state values}
\usage{
steady_states(data, width = 30)
}
\arguments{
\item{data}{data.frame returned from `prepare_for_modeling()`}
}
\value{
data.table
}
\description{
Default width used to calculate steady state is 30 seconds
}
