% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{subj_age<-}
\alias{subj_age<-}
\title{Set the value of `subj_age`}
\usage{
subj_age(x) <- value
}
\arguments{
\item{x}{object}

\item{value}{new value}
}
\value{
modified object
}
\description{
Set the value of `subj_age`
}
