% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{subj_age}
\alias{subj_age}
\title{Acessor for `bads`}
\usage{
subj_age(x)
}
\arguments{
\item{x}{object}
}
\value{
value if `subj_age` exists in the object, NULL otherwise
}
\description{
Acessor for `bads`
}
