% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{subj_height<-}
\alias{subj_height<-}
\title{Set the value of \code{subj_height}}
\usage{
subj_height(x) <- value
}
\arguments{
\item{x}{object}

\item{value}{new value}
}
\value{
modified object
}
\description{
Set the value of \code{subj_height}
}
