% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{subj_height}
\alias{subj_height}
\title{Acessor for `subj_height`}
\usage{
subj_height(x)
}
\arguments{
\item{x}{object}
}
\value{
value if `subj_height` exists in the object, NULL otherwise
}
\description{
Acessor for `subj_height`
}
