% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{subj_id<-}
\alias{subj_id<-}
\title{Set the value of `subj_id`}
\usage{
subj_id(x) <- value
}
\arguments{
\item{x}{object}

\item{value}{new value}
}
\value{
modified object
}
\description{
Set the value of `subj_id`
}
