% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{subj_id}
\alias{subj_id}
\title{Acessor for `subj_id`}
\usage{
subj_id(x)
}
\arguments{
\item{x}{object}
}
\value{
value if `subj_id` exists in the object, NULL otherwise
}
\description{
Acessor for `subj_id`
}
