% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{subj_sex<-}
\alias{subj_sex<-}
\title{Set the value of `subj_sex`}
\usage{
subj_sex(x) <- value
}
\arguments{
\item{x}{object}

\item{value}{new value}
}
\value{
modified object
}
\description{
Set the value of `subj_sex`
}
