% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{subj_sex}
\alias{subj_sex}
\title{Acessor for `subj_sex`}
\usage{
subj_sex(x)
}
\arguments{
\item{x}{object}
}
\value{
value if `subj_sex` exists in the object, NULL otherwise
}
\description{
Acessor for `subj_sex`
}
