% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{subj_weight<-}
\alias{subj_weight<-}
\title{Set the value of \code{subj_weight}}
\usage{
subj_weight(x) <- value
}
\arguments{
\item{x}{object}

\item{value}{new value}
}
\value{
modified object
}
\description{
Set the value of \code{subj_weight}
}
