% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{subj_weight}
\alias{subj_weight}
\title{Acessor for \code{subj_weight}}
\usage{
subj_weight(x)
}
\arguments{
\item{x}{object}
}
\value{
value if \code{subj_weight} exists in the object, NULL otherwise
}
\description{
Acessor for \code{subj_weight}
}
