% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{types_from_names}
\alias{types_from_names}
\title{Label On/Off Kinetics Sections}
\usage{
types_from_names(data, col_types, col_names)
}
\arguments{
\item{data}{data.frame}

\item{col_types}{character vector of col_types}

\item{col_names}{character vector of col_names}
}
\value{
data.frame
}
\description{
Label On/Off Kinetics Sections
}
