% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{unnest_dt}
\alias{unnest_dt}
\title{Unnest data.table}
\usage{
unnest_dt(dt, col, id)
}
\arguments{
\item{dt}{nested data.table}

\item{col}{column to unnest}

\item{id}{id to group by}
}
\value{
unnested data.table
}
\description{
Unnest data.table
}
