% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{col_names}
\alias{col_names}
\title{Acessor for \code{col_names}}
\usage{
col_names(x)
}
\arguments{
\item{x}{object}
}
\value{
value if \code{col_names} exists in the object, NULL otherwise
}
\description{
Acessor for \code{col_names}
}
