% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{col_types}
\alias{col_types}
\title{Acessor for `col_types`}
\usage{
col_types(x)
}
\arguments{
\item{x}{object}
}
\value{
value if `col_types` exists in the object, NULL otherwise
}
\description{
Acessor for `col_types`
}
