% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{crop}
\alias{crop}
\title{Crop data by meas_start and meas_end}
\usage{
crop(data, meas_start = 0, meas_end = NULL)
}
\arguments{
\item{data}{data.frame}

\item{meas_start}{start of testing in seconds. defaults to 0.}

\item{meas_end}{end of testing in seconds. defaults to NULL which does not
remove anything from the end of the data.frame}
}
\value{
data.frame
}
\description{
Crop data by meas_start and meas_end
}
