% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.R
\name{crop}
\alias{crop}
\title{Remove rows from data}
\usage{
crop(dt, first, last, col)
}
\arguments{
\item{dt}{data.frame}

\item{first}{A numeric. If \code{col} is provided, rows with value less than
\code{first} in column \code{col} are removed.  If \code{col} is not provided, rows
with row number less than \code{first} are removed.
Defaults to NULL.}

\item{last}{A numeric. If \code{col} is provided, rows with value greater than
\code{last} in column \code{last} are removed.  If \code{col} is not provided, rows with
row number greater than \code{last} are removed.
Defaults to NULL.}

\item{col}{A character name of column to use to evaluate \code{first} and \code{end}.
Defaults to NULL.

If \code{first},\code{last}, and \code{col} are NULL, the original data will be returned
with a message that no rows were cropped.

\code{first} and \code{last} are not inclusive. If \code{first} = 1, row 1 will not be removed.
To remove row 1, \code{first} must be 2.}
}
\value{
data.frame
}
\description{
Delete rows from data. If \code{col} is provided, works like
\code{dplyr::filter} and filters values in column \code{col} that are
greater than \code{first} and less than \code{last}.
If \code{col} is not provided, works like subsetting based on row numbers.
Rows with row numbers less than \code{first} and greater than \code{last} are removed from
the dataset.
}
