% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{device_site}
\alias{device_site}
\title{Acessor for `device_site`}
\usage{
device_site(x)
}
\arguments{
\item{x}{object}
}
\value{
value if `device_site` exists in the object, NULL otherwise
}
\description{
Acessor for `device_site`
}
