% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggook}
\alias{ggook}
\title{Wrapper around ggplot functions to plot on/off kinetics}
\usage{
ggook(data, xmin = 0, xmax = 960)
}
\arguments{
\item{data}{data frame to use}

\item{xmin}{minimum x value to display}

\item{xmax}{maximum x value to display}
}
\description{
Wrapper around ggplot functions to plot on/off kinetics
}
