% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{last_samp<-}
\alias{last_samp<-}
\title{Setter for \code{last_samp}}
\usage{
last_samp(x) <- value
}
\arguments{
\item{x}{object}

\item{value}{value to set as \code{last_samp}}
}
\value{
modified object
}
\description{
Setter for \code{last_samp}
}
