% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{meas_date}
\alias{meas_date}
\title{Acessor for `meas_date`}
\usage{
meas_date(x)

meas_date(x)
}
\arguments{
\item{x}{object}
}
\value{
value if `meas_date` exists in the object, NULL otherwise

value if `meas_date` exists in the object, NULL otherwise
}
\description{
Acessor for `meas_date`

Acessor for `meas_date`
}
