% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{meas_start<-}
\alias{meas_start<-}
\title{Setter for \code{meas_start}}
\usage{
meas_start(x) <- value
}
\arguments{
\item{x}{object}

\item{value}{value to set as \code{meas_start}}
}
\value{
modified object if \code{meas_start} exists, otherwise NULL
}
\description{
Setter for \code{meas_start}
}
