% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nirsdexata-package.R
\docType{package}
\name{nirsdexata-package}
\alias{nirsdexata}
\alias{nirsdexata-package}
\title{nirsdexata: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
