% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_rawdata}
\alias{plot_rawdata}
\title{Plot the On/Off Kinetics of the provided data set}
\usage{
plot_rawdata(rawdata, type = NULL)
}
\arguments{
\item{rawdata}{rawdata object}

\item{type}{type of data to plot. Default setting will facet the plot
by type.}
}
\value{
ggplot
}
\description{
Plot the On/Off Kinetics of the provided data set
}
