% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{remove_rows}
\alias{remove_rows}
\title{Remove rows from the data}
\usage{
remove_rows(data = data.frame(), remove, col = charact)
}
\arguments{
\item{data}{data.frame}

\item{remove}{vector of data to remove}

\item{col}{column to remove `remove` from}
}
\value{
data.frame
}
\description{
input data must be in long format
and values are replaced with NA
replacing with NA helps with ggplot geom_line functions
}
