% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{steady_states}
\alias{steady_states}
\title{Calculate steady state values}
\usage{
steady_states(dt, width = 30)
}
\arguments{
\item{dt}{data.frame returned from \code{prepare_for_modeling()}}

\item{width}{A double, number of seconds}
}
\value{
data.table
}
\description{
Default width used to calculate steady state is 30 seconds
}
